requirements.txt
aiohttp==3.9.4rc0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
dandroid126-logger @ git+https://github.com/dandroid126/logger@00151aac1802caab2026a0fd2d49510504108297
discord.py==2.3.2
exceptiongroup==1.2.0
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
multidict==6.0.4
packaging==23.2
pluggy==1.4.0
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tomli==2.0.1
tzlocal==5.2
urllib3==2.1.0
yarl==1.9.4
